Rename chrome-version.ts to chrome.ts
